multer

+++ src/services/Profile.js
@@ -1,4 +1,8 @@
 const ProfileRepository = require("../repositories/Profile");
+const fs = require("fs");
+const path = require("path");
+const Profile = require("../models/Profile");
+const DbLog = require("../models/DbLog");
 
 class ProfileService {
   async createProfile(profileData) {
@@ -9,12 +13,43 @@ class ProfileService {
     return await ProfileRepository.read(id);
   }
 
-  async updateProfile(id, profileData) {
-    await ProfileRepository.update(id, profileData);
+  async updateProfile(id, updateData, photoPath) {
+    const profile = await Profile.findByPk(id);
+
+    if (!profile) throw new Error("Профиль не найден");
+
+    // Удалить старое фото, если новое загружено
+    if (photoPath && profile.photo && fs.existsSync(profile.photo)) {
+      fs.unlinkSync(profile.photo);
+    }
+
+    if (photoPath) updateData.photo = photoPath;
+
+    await profile.update(updateData);
+
+    await DbLog.create({
+      operation: "update",
+      entity: "Profile",
+      data: { id, ...updateData },
+    });
   }
 
   async deleteProfile(id) {
+  
+    const profile = await Profile.findByPk(id);
+    if (!profile) throw new Error("Профиль не найден");
+
+    if (profile.photo && fs.existsSync(profile.photo)) {
+      fs.unlinkSync(profile.photo);
+    }
+
     await ProfileRepository.delete(id);
+
+    await DbLog.create({
+      operation: "delete",
+      entity: "Profile",
+      data: { id },
+    });
   }
 }
